+++ pentobi/Comment.qml
@@ -22,8 +22,26 @@ Item {
                                  : Qt.alpha(theme.colorText, 0.1)
     }
     ScrollView {
+        id: scrollView
+
         anchors.fill: parent
-        ScrollBar.vertical.minimumSize: 0.2
+        ScrollBar.vertical.minimumSize: 0.15
+        // Workaround for QTBUG-140033 (Scrollbar not painted in Fusion style,
+        // Qt 6.9.2)
+        ScrollBar.vertical.contentItem: Rectangle {
+            implicitWidth: 6
+            radius: 3
+            color: theme.colorText
+            opacity:
+                if (scrollView.ScrollBar.vertical.pressed)
+                    return 0.4
+                else if (scrollView.ScrollBar.vertical.hovered)
+                    return 0.3
+                else if (scrollView.ScrollBar.vertical.size < 1)
+                    return 0.2
+                else
+                    return 0
+        }
 
         TextArea {
             id: textArea