+++ NEWS.md
@@ -10,6 +10,7 @@ Version 27.0-dev (main)
 * Menu no longer covers menu button in mobile view (if not on Android),
   which works better for Plasma Mobile.
 * Added close link in help viewer
+* Workaround for QTBUG-140033 (Scrollbar not painted in Fusion style)
 
 Version 26.5-dev (v26-fixes)
 ----------------------------